+++ Service/numbers/iAddNum.cs
@@ -2,5 +2,5 @@ namespace Fooa_MiniChallenge2_towtofour.Service.numbers;
 
 public interface iAddNum 
 {
-    public string addNumbers(double num1, double num2);
+    public string addNumbers(int num1, int num2);
 }
